requirements.txt
selenium==4.15.2
openai==1.3.0
pandas==2.1.3
python-dotenv==1.0.0
webdriver-manager==4.0.1
beautifulsoup4==4.12.3
pyautogui==0.9.54
requests==2.31.0
urllib3==2.0.7
undetected-chromedriver==3.5.4